Safari-609.1.11.1/LayoutTests/canvas/philip/tests/size.attributes.reflect.setcontent.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.reflect.setcontent</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.reflect.setcontent.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

canvas.setAttribute('width', '120');
canvas.setAttribute('height', '60');
_assertSame(canvas.getAttribute('width'), '120', "canvas.getAttribute('width')", "'120'");
_assertSame(canvas.getAttribute('height'), '60', "canvas.getAttribute('height')", "'60'");
_assertSame(canvas.width, 120, "canvas.width", "120");
_assertSame(canvas.height, 60, "canvas.height", "60");


});
</script>
