LayoutTests/canvas/philip/tests/size.attributes.reflect.setcontent-expected.txt

< [index] >
size.attributes.reflect.setcontent
Setting content attributes updates IDL and content attributes
References: size.reflect
Actual output:
Expected output:

Passed
